Refactor chart components and update Vico library
- Updated Vico library from version 2.0.10 to 2.1.1.
- Refactored `PrimaryLayer.kt` to use `CartesianLayerRangeProvider.fixed` for setting Y-axis range, removing custom range provider logic.
- Modified `GraphIntent.kt` and `GraphState.kt` to change `markerIndex` type from `Int?` to `Double?` to support more precise marker positioning.
- Updated `GraphViewModel.kt` to adjust Y-axis calculation based on the visible range and implement new logic for handling scroll events and marker updates.
- Refactored `ChartHostSection.kt` to handle chart click events for marker placement and update scroll stopped event handling.
- Removed custom pointer input handling and marker visibility listener from `GraphView.kt`, relying on Vico's built-in click handling.
- Simplified `GraphMarker.kt` by removing `markerIndex` and `xLabels` dependencies from `valueFormatter`.

+++ Android/app/src/main/java/com/dmdbrands/gurus/weight/features/common/components/chart/viewmodel/GraphIntent.kt
@@ -27,7 +27,7 @@ sealed interface GraphIntent : IReducer.Intent {
   data class UpdateTargetRange(val minTarget: Long?, val maxTarget: Long?) : GraphIntent
 
   /** Update marker index */
-  data class UpdateMarkerIndex(val markerIndex: Int?) : GraphIntent
+  data class UpdateMarkerIndex(val markerIndex: Double?) : GraphIntent
 
   data class UpdateSavedTarget(val target: Long) : GraphIntent
